Changed file name, removed test file, and commented unused code to merge
